[sailfish-policy] Import from sailfish-mdm. JB#59977
Import sources as of revision 0.4.16 of the sailfish-mdm package.

If you want to inspect the history of the imported content, fetch the
replace reference first with:

git config --add remote.origin.fetch "+refs/replace/*:refs/replace/*"
git fetch

(replace "origin" with the actual name of your remote)

After that, it will appear as a branch merged with this commit. See
git-replace(1) for more details.